Move WebAuthn components into dedicated folder
