+++ settings/src/components/webauthn/webauthn.js
@@ -8,7 +8,8 @@ import { Icon, check } from '@wordpress/icons';
 /**
  * Internal dependencies
  */
-import { GlobalContext } from '../script';
+import { GlobalContext } from '../../script';
+import ListKeys from './list-keys';
 
 const confirm = window.confirm;
 
@@ -93,38 +94,6 @@ function Manage( { setStep, userKeys } ) {
 				</Button>
 			</p>
 		</>
-	);
-}
-
-/**
- * Render the list of keys.
- *
- * @param props
- * @param props.keys
- */
-function ListKeys( { keys } ) {
-	return (
-		<ul>
-			{ keys.map( ( key ) => (
-				<li key={ key.id }>
-					{ key.name }
-
-					{ /* todo add onclick handler that pops up a <Modal> to confirm. maybe pass in from parent? */ }
-					<Button
-						variant="link"
-						data-id={ key.id }
-						aria-label="Delete"
-						onClick={ () =>
-							confirm(
-								'Modal H4 Remove Key? p Are you sure you want to remove the "" security key? Button Cancel Button Remove Key'
-							)
-						}
-					>
-						Delete
-					</Button>
-				</li>
-			) ) }
-		</ul>
 	);
 }
